resolve #163 per Team input on Mission integration
This PR resolves #163 based on Team input to make ensure close alignment in both directions between the Vision and w3.org/mission, by both pointing to each other and complementing each other. In particular:

Update abstract to be more precise. The Vision does not articulate the Mission, w3.org/mission does that. The Vision does articulate the values underpinning the Mission.

Amended abstract text is directly what was recommended by the Team, and edited accordingly by the editor.

Additional reasoning: remove “it operates” so that “it makes decisions” is stronger on its own, and also because how W3C operates is an entirely distinct concept.

Update purpose of this document to explicitly link to the Mission, and update the sentence doing so with language consistent with the abstract update.

Minor edits to make the Vision more timeless as a self-standing document and properly indicate past work (to read as a candidate for Statement)
* removed "today"
* removed "currently" and editorially reword that sentence in Acknowledgments accordingly.