change Acks notably to editors thank+recognize AB + feedback improved
This is an editor's rewrite of the "notably" acknowledgments to a more accurate editors thanking and recognition of key contributors of the Vision's origin document, AB participants who explicitly supported the Vision, and specific feedback from individuals which improved this document. Minor edit to the adjacent line to link to the broader AB Vision wiki page rather than a fragment of only one year.
This PR resolves issues #270 and #277.